Сделал модуляцию условную
